# Conflicts:
#	src/build/config/config.yaml
#	src/build/help/help.xml
#	src/command/server/ping.c
#	src/command/server/server.c
#	src/common/io/tls/client.c
#	src/common/io/tls/client.h
#	src/common/io/tls/common.c
#	src/common/io/tls/common.h
#	src/common/io/tls/server.c
#	src/common/io/tls/server.h
#	src/config/config.auto.h
#	src/config/parse.auto.c
#	src/protocol/helper.c
#	src/storage/azure/storage.c
#	src/storage/gcs/storage.c
#	src/storage/s3/storage.c
#	test/certificate/README.md
#	test/certificate/pgbackrest-test-client.cnf
#	test/certificate/pgbackrest-test-server.cnf
#	test/src/common/harnessServer.c
#	test/src/common/harnessServer.h
#	test/src/module/command/helpTest.c
#	test/src/module/common/ioTlsTest.c
#	test/src/module/protocol/protocolTest.c
#	test/src/valgrind.suppress.u16

+++ src/common/io/tls/client.c
@@ -385,7 +385,7 @@ tlsClientNew(
         memContextCallbackSet(objMemContext(driver), tlsClientFreeResource, driver);
 
         // Enable safe compatibility options
-        SSL_CTX_set_options(driver->context, (long)SSL_OP_ALL);
+        SSL_CTX_set_options(driver->context, SSL_OP_ALL);
 
         // Set location of CA certificates if the server certificate will be verified
         if (driver->verifyPeer)
